redoing redraw of partial

+++ ctext.h
@@ -345,7 +345,7 @@ class ctext
 		int8_t ob_end();
 
 		int8_t highlight(int32_t row, int32_t col, int32_t len);
-		int8_t redraw_partial_test();
+		int16_t redraw_partial_test();
 
 	private:
 		int8_t y_scroll_calculate(int32_t amount, int32_t *x, int32_t *y);
